# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Epic Poetry Cafe Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Owner dashboard summary
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/profitability:
    get:
      operationId: getItemProfitability
      tags: [dashboard]
      summary: Item profitability
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profitability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemProfitability"
  /dashboard/daily-pl:
    get:
      operationId: getDailyPL
      tags: [dashboard]
      summary: Daily P&L
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily P&L
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPL"
  /dashboard/consumption-variance:
    get:
      operationId: getConsumptionVariance
      tags: [dashboard]
      summary: Ingredient consumption variance
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Variance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionVariance"
  /dashboard/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [dashboard]
      summary: Sales trend
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"
  /dashboard/expense-breakdown:
    get:
      operationId: getExpenseBreakdown
      tags: [dashboard]
      summary: Expense breakdown by category
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expense breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /dashboard/vendor-spend:
    get:
      operationId: getVendorSpendAnalytics
      tags: [dashboard]
      summary: Vendor spend analytics
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendor spend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorSpendItem"
...
components:
  schemas:
...
    TopSellingItem:
      type: object
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: number
        revenue:
          type: number
      required:
        - menuItemId
        - menuItemName
        - quantity
        - revenue
...
    DashboardSummary:
      type: object
      properties:
        todaySales:
          type: number
        todayExpenses:
          type: number
        todayWaste:
          type: number
        todayEstimatedProfit:
          type: number
        mtdSales:
          type: number
        mtdExpenses:
          type: number
        mtdWaste:
          type: number
        mtdProfit:
          type: number
        lowStockCount:
          type: integer
        pendingRecurringExpenses:
          type: integer
        topItemsByRevenue:
          type: array
          items:
            $ref: "#/components/schemas/TopSellingItem"
        topItemsByProfit:
          type: array
          items:
            $ref: "#/components/schemas/ItemProfitability"
        topWasteItems:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
        insights:
          type: array
          items:
            type: string
      required:
        - todaySales
        - todayExpenses
        - todayWaste
        - todayEstimatedProfit
        - mtdSales
        - mtdExpenses
        - mtdWaste
        - mtdProfit
        - lowStockCount
        - topItemsByRevenue
        - topItemsByProfit
        - alerts
        - insights
    AlertItem:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
        message:
          type: string
        relatedId:
          type: ["integer", "null"]
      required:
        - type
        - severity
        - message
    ItemProfitability:
      type: object
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantitySold:
          type: number
        revenue:
          type: number
        productionCost:
          type: number
        grossProfit:
          type: number
        marginPercent:
          type: number
      required:
        - menuItemId
        - menuItemName
        - quantitySold
        - revenue
        - productionCost
        - grossProfit
        - marginPercent
    DailyPL:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        materialCost:
          type: number
        wasteCost:
          type: number
        trialCost:
          type: number
        allocatedExpenses:
          type: number
        allocatedUtilities:
          type: number
        grossProfit:
          type: number
        operatingProfit:
          type: number
        grossMarginPercent:
          type: number
        operatingMarginPercent:
          type: number
      required:
        - date
        - totalSales
        - materialCost
        - wasteCost
        - trialCost
        - allocatedExpenses
        - grossProfit
        - operatingProfit
    ConsumptionVariance:
      type: object
      properties:
        ingredientId:
          type: integer
        ingredientName:
          type: string
        actualConsumed:
          type: number
        theoreticalConsumed:
          type: number
        variance:
          type: number
        variancePercent:
          type: number
        uom:
          type: string
      required:
        - ingredientId
        - ingredientName
        - actualConsumed
        - theoreticalConsumed
        - variance
        - variancePercent
        - uom
    TrendDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: ["string", "null"]
      required:
        - date
        - value
    CategoryBreakdown:
      type: object
      properties:
        name:
          type: string
        value:
          type: number
        count:
          type: integer
      required:
        - name
        - value
    VendorSpendItem:
      type: object
      properties:
        vendorId:
          type: integer
        vendorName:
          type: string
        totalSpend:
          type: number
        purchaseCount:
          type: integer
      required:
        - vendorId
        - vendorName
        - totalSpend
        - purchaseCount
...